clear warning

+++ Elevator/ElevatorClass.cpp
@@ -70,5 +70,5 @@ bool Elevator::isFull()
 }
 int Elevator::userCount()
 {
-	return users.size();
+	return (int)users.size();
 }
